requirements.txt
google-cloud-bigquery
pandas
db-dtypes
matplotlib
seaborn
jupyter
ipykernel
pyarrow
scikit-learn
xgboost
joblib